initial commit
First version*, when it was still called "sehtml" (S-Expression based HTML dialect), extremely simplistic and not what I am aiming for. Code as of 2022-01-10 and it is very ugly. I later rewrote most of it, as it was too ugly to update (or even look at), and too hard to maintain. Published, because I wanted to preserve it somewhere, yet not sure why though lol

*actually it is the second version, the first one I wrote a few years ago, but I lost the code, so let us consider this one as the first